pyproject.toml
[project]
name = "rent-vs-own"
version = "0.1.0"
description = "Compare long-term economic gains from renting versus owning using ACS and HMDA data."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.32.0",
    "google-cloud-bigquery>=3.23.0",
    "typer>=0.9.0",
]

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
rent-vs-own = "rent_vs_own.cli:app"
